Wait until the service is available.
We would wait until AWS EC2 services return a status of “running” and then
immediately return an IP address, but at this time the sshd service is not
running yet.

So in this patch we add a small TCP ping-like facility that tries to connect
to the TCP port 22 in a busy wait loop, and we only return to the caller
when we could connect.

With that the Makefile seems to work reliably.