Rewrite parameters handling.
The parameters of the benchmark should not be hard coded into the Makefiles
but rather come from a single tpch.ini file. This commit makes that happen,
and then review the Python tpch.py driver code to use this ini file and run
the `make Makefile.loader` with the right set of parameters.