Implement support for PostgreSQL and Citus.
At long last, add support for those systems that are not to be managed by
the infra.py utility script. We only expect a connection string (DSN) to the
database services here, so that it's easy to benchmark Citus Cloud, our real
target in this exercise.

Also, review the README.md documentation to make it worthwhile.