Use logging rather than print().
The idea is that the benchmark will be running for a long time (several
hours), and we can't depend on the fact that the controller is connected
through ssh for the whole duration of the tests. So we will need to detach
it somehow, and we can't then log to stdout as we do.

In passing, the responsibility of the progress report log message need to
stay in the caller to the pooling facilities, and the way to make that
happen in Python is either to give yet another function as an argument or to
build a class with an abstract method, and implement the method in the
subclass. That's what we do here.