Improve the tracking system.
Also remame the “result” Python module into “tracking”, which is a more
proper name given that it only registers/track results anyway.

Simplify the tracking SQL schema and add a view to make it easier to have a
quick look at what a run is made off, seeing the timings and the number of
queries executed etc.