Run the benchmark as a managed service.
We don't really need a daemon/service infrastructure here, something
interactive only would be plenty good enough. But we need to be resilient to
loosing an ssh connection in between the controller and the loader nodes.

To that end, we use a Service Management API and run our process in the
background, and the Makefile is now redirecting the output with the help of
/usr/bin/tee so that we can tail -f later, when disconnected.