Implement --detach.
The idea of the managed service in the previous patch doesn't solve the
problem we have. What we need is to run as a Unix daemon, which despite its
name the lib python-daemon doesn't implement. The Python lib “service” is
using python-daemon and is the one to use here.

Once the code might run as a service, we want to log to a file of course,
and more than that, the environement variables tricks stop working
automatically for us. Implement some efforts to pass-in the variables down
the forks(), and also register some PATH at “compile/load” time.

In passing, clean-up the code to PEP-8 standards.

Oh, and remove the schema/README.md, there's not that much that comes from
Tomas' repository anymore, only the bits that came with TPCH anyway.