Register vacuum operations as jobs.
Now that vacuum is handled in a schema file, just register it in our log
trail as another schema job, with its own entry.

In passing, fix a bug in the tracking schema where we couldn't run the same
benchmark job more than once: that's something we want to be able to do!