Implement recursive schedules and parallel schedules.
Recursive schedule is a pretty simple idea: a schedule consists of either
jobs or schedules, which are defined at the same level. That's easy enough,
and necessary to be able to have some flexibility in editing the ini files.

Parallel schedules were supposed to be really easy given that we already
have tpch.task_dist after all, but it turns out we can't use that here.
Also turns out we need to use logging.handlers.QueueHandler() and some more
infrastructure, as described in the following document:

  https://docs.python.org/3/howto/logging-cookbook.html#logging-to-a-single-file-from-multiple-processes

Anyway, I guess that's done now.