Review benchmark setup and parameters.
Turns out the c5.4xlarge was crawling under the load with 16 concurrent
processes doing DGBEN, because each is using 2 to 3 GB of RAM and we got
some OOM killer involved. Switch to c5.18xlarge and fix some other bugs in
the setup.

One of those bugs is about foreign-keys. To have numbers we can compare,
don't install them on the RDS/Aurora tests either.