Fix memory leak in TPCH direct load support.
I've not been very careful in how I used the PQExpBuffer API and as a result
we were leaking all the allocated memory.

The code is using a single stack allocated entry for each object's data, so
that there's no memory allocation during the load. Then the printer uses
fprintf() and doesn't allocate buffers. We allocate a buffer in the COPY
support, so we need to take care of memory correctly.

In this patch we both fix the memory usage (see tpch-pg/src/dss.h), review
the function prototypes to simplify them thanks to using the static
initialisation of the connection variable to track if we need to connect to
PostgreSQL.

This could arguably have been two different patches…