Implement a TPC-H controller in Python.
This allows to pass forward some of the limitations in the Makefile based
approach, namely the following two hurdles:

  - running several benchmarks in parallel,
  - managing benchmark targetting a sublist of the expected 4 systems.

The new system still keeps AWS outputs in local files, they are stored in a
per-benchmark directory now (e.g. aws.out/engage_busily/{aws,logs,results}).
Given such a structure, the tpch.ini and infra.ini files are kept in that
same area for the duration of the test, so that you can fiddle with the main
ones and still have a good reference of what's running where.

The fetch-logs, dump-results, and merge-results targets are implemented to
in the ./control.py update command.