Add --system to most of the control.py commands.
This allows to control a single system from a benchmark run. It comes in
handy when there's a spurious error on one system and you want to restart
the benchmark on this system, without altering the other ones.

To have a single system be restarted on the loader, you could now:

  ./control.py cancel --system foo benchmark_name
  ./control.py benchmark --system foo benchmark_name