Automate creation of the tpch-results database.
Report failure to connect (most probably because the database doesn't exists
yet) at the proper timing, and only once. When the database exists and we
can connect, quickly check that we have our expected tables and views and if
not, run schema/tracking.sql on the result database.

This should make it easier to get started for every user, including those
who are shared results with. ./control.py merge_results --db should just
work for them!