Generalize the Infrastructure systems list.
The way the system names participating in the infrastructure were hard coded
meant it was not possible to run the same benchmark with several Citus
clusters or RDS 9.6 and 10.1, for instance. This patch makes it possible.

The infra.ini setup sections now have a “type” option allowing for several
sections of the same type with different names. When the type is not
specified then we default to the previous spelling and hard-coded types: the
section name must then be one of 'rds', 'aurora', 'citus' or 'pgsql'.