Only parse timing output from SELECT queries.
Some TPC-H queries contain a series of CREATE VIEW, SELECT, DROP VIEW, and
the time parsing was naive enough to then try to register 3 timings for a
single query.

This patch makes it only a tad less naive and only registers the timing of
running the intermediate SELECT query, discarding the create/drop of the
view.

Without this patch, we get the following traceback, that happens because we
have more timings than queries:

  File "/Users/dim/dev/CitusData/tpch/tpch/run/utils.py", line 65, in parse_psql_timings
    timings[qstream[i]] = timing
IndexError: list index out of range