Implement support for EC2 based PostgreSQL instances.
This allows testing core PostgreSQL packages from apt.postgresql.org running
in an Amazon Ubuntu image. We use an EC2 instance and provision the database
service ourselves.

With that it's now possible to run TPC-H against several versions of
PostgreSQL, adding to our support for Citus Cloud, RDS, and Aurora.
Supported PostgreSQL version are now 9.6 and 10, and 11 is on its way.