Fix infra environment.
In particular the current version of python-daemon in pipy is broken, so we
need to make sure to use the previous one. In passing, make it so that any
execute_remote_command() call that receives a non-zero return code raises a
RuntimeError, so that we don't try the next step when the current one has
failed already.