Add support for custom shared_buffers and pg_prewarm.
It was supposed to help with the JIT tests, currently to no avail. At least
it's possible to force the TPCH data in shared_buffers when it can fit, as
in the plans/llvm-q1-schedule.ini where we have 18GB of data and indexes and
a machine with 30GB of RAM. We then use 20GB of shared_buffers.